Реализовал возможность использование опции plain: для рендера обычного текста
