+++ lib/simpler/render_types/erb.rb
@@ -1,0 +1,37 @@
+require 'erb'
+
+module Simpler
+  class Erb
+    VIEW_BASE_PATH = 'app/views'.freeze
+
+    def initialize(env)
+      @env = env
+    end
+
+    def render(binding)
+      template = File.read(template_path)
+
+      ERB.new(template).result(binding)
+    end
+
+    private
+
+    def controller
+      @env['simpler.controller']
+    end
+
+    def action
+      @env['simpler.action']
+    end
+
+    def template
+      @env['simpler.template']
+    end
+
+    def template_path
+      path = template || [controller.name, action].join('/')
+
+      Simpler.root.join(VIEW_BASE_PATH, "#{path}.html.erb")
+    end
+  end
+end
