+++ lib/simpler/render_types/plain.rb
@@ -1,0 +1,11 @@
+module Simpler
+  class Plain
+    def initialize(env)
+      @env = env
+    end
+
+    def render(_binding)
+      @env['simpler.plain_text']
+    end
+  end
+end
