+++ lib/simpler/view.rb
@@ -1,39 +1,20 @@
-require 'erb'
+require_relative 'render_types/erb'
+require_relative 'render_types/plain'
 
 module Simpler
   class View
 
-    VIEW_BASE_PATH = 'app/views'.freeze
+    VALID_FORMATS = {'erb, '}
+
+    DEFAULT_RENDER_TYPE = 'erb'.freeze
 
     def initialize(env)
       @env = env
     end
 
     def render(binding)
-      template = File.read(template_path)
-
-      ERB.new(template).result(binding)
+      render_type = @env['simpler.render_type'] || DEFAULT_RENDER_TYPE
+      Simpler.const_get(render_type.capitalize).new(@env).render(binding)
     end
-
-    private
-
-    def controller
-      @env['simpler.controller']
-    end
-
-    def action
-      @env['simpler.action']
-    end
-
-    def template
-      @env['simpler.template']
-    end
-
-    def template_path
-      path = template || [controller.name, action].join('/')
-
-      Simpler.root.join(VIEW_BASE_PATH, "#{path}.html.erb")
-    end
-
   end
 end
